chapter_004/src/main/java/ru/job4j/lambda/FindFunctionValue.java
package ru.job4j.lambda;

import java.util.ArrayList;
import java.util.List;
import java.util.function.Function;

public class FindFunctionValue {
    public static List<Double> diapason(int start, int end, Function<Double, Double> func) {
        List<Double> result = new ArrayList<>();
        for (int i = start; i < end; i++) {
            result.add(func.apply((double) i));
        }
        return result;
    }

    public static void main(String[] args) {
        System.out.println(diapason(2, 5, x -> 2 * x + 1));
    }
}
